Fix some warnings

+++ OpenGL/src/VertexArray.cpp
@@ -4,36 +4,36 @@
 
 VertexArray::VertexArray()
 {
-	GLCall(glGenVertexArrays(1, &m_RendererID));
+    GLCall(glGenVertexArrays(1, &m_RendererID));
 }
 
 VertexArray::~VertexArray()
 {
-	GLCall(glDeleteVertexArrays(1, &m_RendererID));
+    GLCall(glDeleteVertexArrays(1, &m_RendererID));
 }
 
 void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout& layout)
 {
-	Bind();		// Bind the array
-	vb.Bind();	// Bind the buffer
-	// Set the layout
-	const auto& elements = layout.GetElements();
-	unsigned int offset = 0;
-	for (unsigned int i = 0; i < elements.size(); i++)
-	{
-		const auto& element = elements[i];
-		GLCall(glEnableVertexAttribArray(i));
-		GLCall(glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), reinterpret_cast<const void*>(offset)));
-		offset += element.count * VertexBufferElement::GetSizeOfType(element.type);
-	}
+    Bind();		// Bind the array
+    vb.Bind();	// Bind the buffer
+    // Set the layout
+    const auto& elements = layout.GetElements();
+    uint64_t offset = 0;
+    for (uint32_t i = 0; i < elements.size(); i++)
+    {
+        const auto& element = elements[i];
+        GLCall(glEnableVertexAttribArray(i));
+        GLCall(glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), reinterpret_cast<const void*>(offset)));
+        offset += element.count * static_cast<uint64_t>(VertexBufferElement::GetSizeOfType(element.type));
+    }
 }
 
 void VertexArray::Bind() const
 {
-	GLCall(glBindVertexArray(m_RendererID));
+    GLCall(glBindVertexArray(m_RendererID));
 }
 
 void VertexArray::Unbind() const
 {
-	GLCall(glBindVertexArray(0));
+    GLCall(glBindVertexArray(0));
 }
